Divided the home page into components
