+++ task-mgmt-module/src/HomePage.jsx
@@ -1,13 +1,49 @@
 import React, { useState, useEffect } from "react";
-import { Link } from "react-router-dom";
-import { Layout, Typography, Card, Button, Badge, Row, Col, Divider, Switch, Space } from "antd";
+import { Layout, Typography, Row, Col, Divider, Switch, Space, Badge } from "antd";
+import { ThemeProvider, useTheme } from "./ThemeContext";
+import ModuleCard from "./ModuleCard";
 
 const { Header, Content, Footer } = Layout;
 const { Title, Text, Paragraph } = Typography;
 
-const HomePage = () => {
-  const [theme, setTheme] = useState('light');
-  const [activeViewId, setActiveViewId] = useState(null);
+// Styles for the animations
+const GlobalStyles = () => {
+  const { themeValues } = useTheme();
+  
+  return (
+    <style>{`
+      @keyframes fadeIn {
+        from { opacity: 0; transform: translateY(20px); }
+        to { opacity: 1; transform: translateY(0); }
+      }
+      
+      .module-card {
+        transition: all 0.3s ease;
+      }
+      
+      .module-card:hover {
+        transform: translateY(-4px);
+        box-shadow: ${themeValues.theme === 'light' 
+          ? '0 12px 24px rgba(0, 0, 0, 0.10)' 
+          : '0 12px 24px rgba(0, 0, 0, 0.35)'} !important;
+      }
+      
+      .module-card:hover .card-image {
+        transform: scale(1.05);
+      }
+      
+      /* Ensure no horizontal scrolling on any screen */
+      body {
+        overflow-x: hidden;
+        max-width: 100vw;
+      }
+    `}</style>
+  );
+};
+
+// Main component content
+const HomePageContent = () => {
+  const { theme, toggleTheme, themeValues } = useTheme();
   const [loaded, setLoaded] = useState(false);
 
   const mockImages = {
@@ -52,80 +88,8 @@ const HomePage = () => {
 
   useEffect(() => {
     setLoaded(true);
-    return () => {
-      const canvas = document.getElementById('particles-canvas');
-      if (canvas) canvas.remove();
-    };
   }, []);
-
-  useEffect(() => {
-    if (loaded) {
-      const canvas = document.createElement('canvas');
-      canvas.id = 'particles-canvas';
-      canvas.style.position = 'fixed';
-      canvas.style.top = '0';
-      canvas.style.left = '0';
-      canvas.style.width = '100%';
-      canvas.style.height = '100%';
-      canvas.style.pointerEvents = 'none';
-      canvas.style.zIndex = '1';
-      document.body.appendChild(canvas);
-
-      const ctx = canvas.getContext('2d');
-      const particles = [];
-      const particleCount = 30; 
-
-      // Set canvas dimensions to match viewport
-      canvas.width = window.innerWidth;
-      canvas.height = window.innerHeight;
-
-      for (let i = 0; i < particleCount; i++) {
-        particles.push({
-          x: Math.random() * canvas.width,
-          y: Math.random() * canvas.height,
-          radius: Math.random() * 2.5 + 0.5, 
-          color: theme === 'light' ? 'rgba(24, 144, 255, 0.12)' : 'rgba(209, 213, 219, 0.08)',
-          speedX: Math.random() * 0.25 - 0.125,
-          speedY: Math.random() * 0.25 - 0.125
-        });
-      }
-
-      const animate = () => {
-        requestAnimationFrame(animate);
-        ctx.clearRect(0, 0, canvas.width, canvas.height);
-        
-        particles.forEach(particle => {
-          ctx.beginPath();
-          ctx.arc(particle.x, particle.y, particle.radius, 0, Math.PI * 2);
-          ctx.fillStyle = particle.color;
-          ctx.fill();
-          
-          particle.x += particle.speedX;
-          particle.y += particle.speedY;
-          
-          if (particle.x < 0) particle.x = canvas.width;
-          if (particle.x > canvas.width) particle.x = 0;
-          if (particle.y < 0) particle.y = canvas.height;
-          if (particle.y > canvas.height) particle.y = 0;
-        });
-      };
-      
-      animate();
-      
-      const handleResize = () => {
-        canvas.width = window.innerWidth;
-        canvas.height = window.innerHeight;
-      };
-      
-      window.addEventListener('resize', handleResize);
-      return () => window.removeEventListener('resize', handleResize);
-    }
-  }, [loaded, theme]);
-
-  const toggleTheme = () => {
-    setTheme(theme === 'light' ? 'dark' : 'light');
-  };
-
+  
   return (
     <Layout
       className="layout"
@@ -133,19 +97,15 @@ const HomePage = () => {
         minHeight: '100vh',
         maxWidth: '100vw',
         overflowX: 'hidden',
-        background: theme === 'light' 
-          ? 'linear-gradient(to bottom right, #f0f5ff, #ffffff, #f0f7ff)' 
-          : 'linear-gradient(to bottom right, #141414, #1f1f1f, #141414)'
+        background: themeValues.background
       }}
     >
       <Header
         style={{
-          background: theme === 'light' 
-            ? 'rgba(255, 255, 255, 0.95)' 
-            : 'rgba(22, 22, 22, 0.95)',
+          background: themeValues.headerBackground,
           backdropFilter: 'blur(8px)',
           WebkitBackdropFilter: 'blur(8px)',
-          boxShadow: theme === 'light' ? '0 2px 8px rgba(0,0,0,0.06)' : '0 2px 8px rgba(0,0,0,0.2)',
+          boxShadow: themeValues.headerShadow,
           padding: '0 24px',
           position: 'sticky',
           top: 0,
@@ -160,18 +120,18 @@ const HomePage = () => {
         <div style={{ display: 'flex', alignItems: 'center' }}>
           <Text strong style={{ 
             fontSize: '18px', 
-            color: theme === 'light' ? '#1890ff' : '#69c0ff' 
+            color: themeValues.accentColor
           }}>
             CARS24
           </Text>
           <Divider type="vertical" style={{ 
             height: '24px', 
             margin: '0 16px',
-            background: theme === 'light' ? '#e5e7eb' : '#303030' 
+            background: themeValues.dividerColor
           }} />
           <Text strong style={{ 
             fontSize: '18px', 
-            color: theme === 'light' ? 'rgba(0, 0, 0, 0.85)' : 'rgba(255, 255, 255, 0.85)' 
+            color: themeValues.primaryText
           }}>
             Dashboard
           </Text>
@@ -180,7 +140,7 @@ const HomePage = () => {
         <div style={{ display: 'flex', alignItems: 'center', gap: '20px' }}>
           <Text style={{ 
             fontSize: '14px',
-            color: theme === 'light' ? 'rgba(0, 0, 0, 0.45)' : 'rgba(255, 255, 255, 0.45)'
+            color: themeValues.tertiaryText
           }}>
             Task Management Module
           </Text>
@@ -210,13 +170,13 @@ const HomePage = () => {
       >
         <div style={{ textAlign: 'center', marginBottom: '48px' }}>
           <Title level={2} style={{ 
-            color: theme === 'light' ? 'rgba(0, 0, 0, 0.85)' : 'rgba(255, 255, 255, 0.85)',
+            color: themeValues.primaryText,
             marginBottom: '8px'
           }}>
             Module Selection
           </Title>
           <Paragraph style={{ 
-            color: theme === 'light' ? 'rgba(0, 0, 0, 0.65)' : 'rgba(255, 255, 255, 0.65)',
+            color: themeValues.secondaryText,
             maxWidth: '700px',
             margin: '0 auto'
           }}>
@@ -226,152 +186,15 @@ const HomePage = () => {
 
         <Row gutter={[24, 24]}>
           {views.map((view, index) => (
-            <Col xs={24} md={12} lg={8} key={view.id}>
-              <div
-                className="card-container"
-                style={{ 
-                  opacity: 0,
-                  animation: `fadeIn 0.5s ease-out ${0.1 + index * 0.1}s forwards`,
-                  height: '100%'
-                }}
-                onMouseEnter={() => setActiveViewId(view.id)}
-                onMouseLeave={() => setActiveViewId(null)}
-              >
-                <Link to={view.path} style={{ textDecoration: 'none' }}>
-                  <Card 
-                    hoverable
-                    className="module-card"
-                    style={{ 
-                      overflow: 'hidden',
-                      height: '100%',
-                      background: theme === 'light' ? '#fff' : '#141414',
-                      borderColor: theme === 'light' ? '#f0f0f0' : '#303030',
-                      boxShadow: theme === 'light' ? '0 4px 12px rgba(0, 0, 0, 0.05)' : '0 4px 12px rgba(0, 0, 0, 0.2)'
-                    }}
-                    bodyStyle={{ padding: '20px' }}
-                    cover={
-                      <div style={{ position: 'relative', height: '192px' }}>
-                        <div style={{ 
-                          position: 'absolute', 
-                          inset: 0, 
-                          background: 'linear-gradient(to top, rgba(0,0,0,0.7), rgba(0,0,0,0.4), transparent)',
-                          zIndex: 1 
-                        }} />
-                        
-                        <div style={{ position: 'absolute', top: '12px', right: '12px', zIndex: 2 }}>
-                          <Badge 
-                            count={view.badgeText} 
-                            style={{ 
-                              backgroundColor: view.color,
-                              fontSize: '12px',
-                              fontWeight: 500
-                            }} 
-                          />
-                        </div>
-                        
-                        <img
-                          alt={view.name}
-                          src={view.image}
-                          style={{ 
-                            width: '100%', 
-                            height: '100%', 
-                            objectFit: 'cover',
-                            transition: 'transform 0.7s ease'
-                          }}
-                          className="card-image"
-                        />
-                        
-                        <div style={{ 
-                          position: 'absolute', 
-                          bottom: '16px', 
-                          left: '16px', 
-                          zIndex: 2 
-                        }}>
-                          <Text strong style={{ 
-                            color: '#fff', 
-                            fontSize: '20px',
-                            textShadow: '0 1px 3px rgba(0,0,0,0.3)'
-                          }}>
-                            {view.id === 3 && activeViewId === 3 ? view.hoverName : view.name}
-                          </Text>
-                        </div>
-                      </div>
-                    }
-                  >
-                    <div style={{ 
-                      display: 'flex', 
-                      justifyContent: 'space-between', 
-                      alignItems: 'flex-start',
-                      marginBottom: '12px'
-                    }}>
-                      <div>
-                        <Title level={4} style={{ 
-                          margin: '0 0 4px 0',
-                          color: theme === 'light' ? 'rgba(0, 0, 0, 0.85)' : 'rgba(255, 255, 255, 0.85)'
-                        }}>
-                          {view.id === 3 && activeViewId === 3 ? view.hoverName : view.name}
-                        </Title>
-                        <Paragraph style={{ 
-                          margin: 0,
-                          color: theme === 'light' ? 'rgba(0, 0, 0, 0.65)' : 'rgba(255, 255, 255, 0.65)',
-                          fontSize: '14px'
-                        }}>
-                          {view.description}
-                        </Paragraph>
-                      </div>
-                      <div style={{ 
-                        display: 'flex',
-                        alignItems: 'center',
-                        justifyContent: 'center',
-                        width: '40px',
-                        height: '40px',
-                        borderRadius: '50%',
-                        background: `${view.color}20`,
-                        color: view.color,
-                        fontSize: '20px'
-                      }}>
-                        {view.icon}
-                      </div>
-                    </div>
-                    
-                    <Divider style={{ 
-                      margin: '16px 0',
-                      borderColor: theme === 'light' ? '#f0f0f0' : '#303030'
-                    }} />
-                    
-                    <div style={{ 
-                      display: 'flex', 
-                      justifyContent: 'space-between', 
-                      alignItems: 'center'
-                    }}>
-                      <Text style={{ 
-                        fontSize: '12px',
-                        color: theme === 'light' ? 'rgba(0, 0, 0, 0.45)' : 'rgba(255, 255, 255, 0.45)'
-                      }}>
-                        
-                      </Text>
-                      <Button 
-                        type="primary"
-                        style={{ 
-                          background: view.color,
-                          borderColor: view.color
-                        }}
-                      >
-                        View
-                      </Button>
-                    </div>
-                  </Card>
-                </Link>
-              </div>
-            </Col>
+            <ModuleCard key={view.id} view={view} index={index} />
           ))}
         </Row>
       </Content>
 
       <Footer 
         style={{ 
-          background: theme === 'light' ? '#fff' : '#141414',
-          borderTop: `1px solid ${theme === 'light' ? '#f0f0f0' : '#303030'}`,
+          background: themeValues.cardBackground,
+          borderTop: `1px solid ${themeValues.dividerColor}`,
           maxWidth: '100%',
           padding: '24px'
         }}
@@ -379,13 +202,13 @@ const HomePage = () => {
         <Row gutter={[32, 24]} style={{ maxWidth: '1440px', margin: '0 auto' }}>
           <Col xs={24} md={8}>
             <Title level={4} style={{ 
-              color: theme === 'light' ? 'rgba(0, 0, 0, 0.85)' : 'rgba(255, 255, 255, 0.85)',
+              color: themeValues.primaryText,
               marginBottom: '16px'
             }}>
               CARS24 CFSPL
             </Title>
             <Paragraph style={{ 
-              color: theme === 'light' ? 'rgba(0, 0, 0, 0.65)' : 'rgba(255, 255, 255, 0.65)',
+              color: themeValues.secondaryText,
               fontSize: '14px'
             }}>
               Transforming the way India buys and sells cars, providing technology-driven solutions for a seamless experience.
@@ -394,7 +217,7 @@ const HomePage = () => {
           
           <Col xs={24} md={8}>
             <Title level={5} style={{ 
-              color: theme === 'light' ? 'rgba(0, 0, 0, 0.75)' : 'rgba(255, 255, 255, 0.75)',
+              color: themeValues.primaryText,
               marginBottom: '16px'
             }}>
               
@@ -423,7 +246,7 @@ const HomePage = () => {
               />
             </Space>
             <Paragraph style={{ 
-              color: theme === 'light' ? 'rgba(0, 0, 0, 0.45)' : 'rgba(255, 255, 255, 0.45)',
+              color: themeValues.tertiaryText,
               fontSize: '14px'
             }}>
               ©️ 2025 CARS24 CFSPL. All rights reserved.
@@ -431,35 +254,17 @@ const HomePage = () => {
           </Col>
         </Row>
       </Footer>
-
-      <style>{`
-        @keyframes fadeIn {
-          from { opacity: 0; transform: translateY(20px); }
-          to { opacity: 1; transform: translateY(0); }
-        }
-        
-        .module-card {
-          transition: all 0.3s ease;
-        }
-        
-        .module-card:hover {
-          transform: translateY(-4px);
-          box-shadow: ${theme === 'light' 
-            ? '0 12px 24px rgba(0, 0, 0, 0.10)' 
-            : '0 12px 24px rgba(0, 0, 0, 0.35)'} !important;
-        }
-        
-        .module-card:hover .card-image {
-          transform: scale(1.05);
-        }
-        
-        /* Ensure no horizontal scrolling on any screen */
-        body {
-          overflow-x: hidden;
-          max-width: 100vw;
-        }
-      `}</style>
+      <GlobalStyles />
     </Layout>
+  );
+};
+
+// Wrapped with ThemeProvider
+const HomePage = () => {
+  return (
+    <ThemeProvider>
+      <HomePageContent />
+    </ThemeProvider>
   );
 };
